structure: AnimWorkerObj::m_1c is an act ID (i32), not a void* int|ptr union
Proven: the LOGIC_WORKER_PUMP switch dispatches on it as a small tag (0/0x1d/0x1e/
0x50..0x53/0x3e8/0x1c), CTypeCollRuntime resolves it as an INDEX (IndexToPtr), and
it is never dereferenced anywhere in the tree. CUserLogic::m_prevAnimSetNode and
CGameObject::NotifyHooked's arg follow it; the g_buteTree act lookups route through
the existing ActFindId/ActInsertId seam. GetNameRecord(s)/ScratchResolve take i32.

Costs 0.27 fuzzy / 7 exacts to /O2 register-class shifts (MAX 75.7368 preserved,
config/match-max.tsv). 2022 -> 2013.

+++ src/Bute/SymTab.cpp
@@ -444,7 +444,7 @@ i32 CRezDirNode::Load(i32 childFlag) {
     }
 
     RezSrc* src = m_src;
-    if (src->m_8 == 0 || static_cast<u32>(src->m_1c) > 1) {
+    if (src->m_8 == 0 || src->m_1c > 1) {
         RezAssertFail("CRezDir::Load Failed! (File is not sorted!)");
         return 0;
     }
